added quit function, cleaned up log output
